Use default thread locking for container image prepare
Process based locking was added to improve container prepare
performance for python2/mistral, according to this change[1] so it
might not be appropriate to use it in an ansible module.

This change removes the ProcessLock so the image preparation uses the
default ThreadLock (and the thread executor).

It is possible that authentication caching doesn't work at all with
the process based executor, and support for that executor should be
removed entirely now that it is not required.

[1]https://opendev.org/openstack/tripleo-common/commit/60afc0eec44f698dd95d9d6ec80dad94a4b07329
Resolves: rhbz#2165527

Change-Id: Ib127b11cc26944dc4c348d7c8b6938bbea0a134a