Deployed Ceph time synchronization
During 'openstack overcloud deploy ceph', there are chances for
apply_spec operation to fail when the chronyd ntp service is down.

Currently cephadm installs chronyd during bootstrap in
command_prepare_host().

This patch introduces the new playbook 'ceph-chrony.yaml' which
makes the chrony available before cephadm bootstrap.
Cephadm will no do it anymore as the bootstrap is now called with
--skip-prepare-host.

The goal here is to use TripleO parameters to configure chrony so that
it is configured consistently.

Resolves: rhbz#2172063
Change-Id: I5d4a6a2f141582c18846057237b0858becc6dacd