Create the snapshots backed with the lv_thinpool pool
Due to the RFE: https://bugzilla.redhat.com/show_bug.cgi?id=2066349, the pool lv_thinpool has been added to the image to create the snapshots backed with this pool. Thus it is not needed to maintain free space in the volume group for the snapshot and revert functionality, and the condition of the 8GB of space is not required.

Delete references to lv_snapshot since it is no longer present.

Change-Id: I6c6341a1d27c9964af7460886789564bdea4066d
(cherry picked from commit d8349341526d908d4cd1061b3f71f63dba8b406d)