Stop using ceph/daemon entrypoint
Due to the recent Ceph 6 (Quincy) container image refactoring,
the old entrypoint is no longer available.
However, the ceph-nfs systemd unit should support both Pacific
and Quincy releases.
This patch introduces a task that is able to detect the version
provided by the ceph container in the Heat stack and renders the
systemd unit accordingly.

Change-Id: I34537db460ff5aaba93da7e104262e26e64cbb1a