Always recreate puppet containers
Podman module just start an existing container if the container already
exists and its definition is not changed. This causes a problem with
puppet containers because detach: false does not take effect when
restarting an existing container, which makes the return code from
puppet ignored.

This restores the previous behavior to always recreated puppet
containers so that return codes from puppet executions are evaluated.

Resolves: rhbz#2185163
Change-Id: Iafa2a4399edad51770247dc76e6871e22022055d