Set grafana VIP to the same CephDashboard network
We have the ability to expose the CephDashboard in networks different
than ctlplane and StorageDashboard (the dedicated composable network).
For this reason we need to make sure the grafana_vip set in the Ceph
cluster by ceph-ansible [1] points to the same network where the
CephDashboard is exposed.
This patch fixes the logic that sets the grafana vip parameter.

Change-Id: I9d3ce30ab0a753af056bb826c093e4f15fe3b07a