Do not use image digest when the Ceph cluster is deployed
We recently hit issues when cephadm applies the spec and tries
to deploy OSDs on the CephStorage nodes.
In particular if a registry != than the undercloud is used,
cephadm might have issues pulling the container images through
podman by digests.
This patch ensures that "use_repo_digest" is set to "false" and
the default container image (identified by its tag) is passed
as part of the config.

Change-Id: I93be367e22d8fe288b1de171d5f8f7e1fde8ae88