Designate: add extra bind IPs to the allow notify list
The designate worker running on hosts that have both the framework
supplied external IPs and the framework deployed bind server
unexpectedly use the special bind IP as a source address. This causes
notify requests from the worker to be rejected. This patch adds the
extra IPs to the allow-notify clause to allow these requests to
succeed.

Patch also adds a missing restart clause for changes to the bind
configuration to take effect.

Closes-bug: #2027608
Depends-On: I0cc4c91e4708221fc429a8f760f08830fe6db220

Change-Id: I503ccceb5942f26a9374b36bfea42a3cad86e4a8