Fix ceph orch command stuck issue
During cephadm playbook execution (as part of stack update)post
FFU upgrade, 'ceph orch status' command is stuck at [1].

This patch will exit the command execution after a timeout and
restarts the current active mgr to fix it.

This fix is just a WA in tripleo until the actual Ceph issue is resolved.

[1] https://opendev.org/openstack/tripleo-ansible/src/branch/master/tripleo_ansible/roles/tripleo_cephadm/tasks/post.yaml#L21

Resolves: rhbz#2222589
Change-Id: I9d9df679467ef1cbd9ad38edf1e7a26a4e6dcdec