Add option to only update a new key on Overcloud
In cases where there are multiple stacks for the overcloud, as in
multicell environments, rotation of the undercloud keys need to be
executed in 2 or more steps.
Currently when a rotation is executed for the first overcloud
stack, it'll fail on the second stack and, also, the underlcoud
will lose connectivity to the others overcloud.
This patch allow the key rotation to be executed in multiple steps
to properly rotate the keys on all overcloud stack.

Change-Id: I186357ccd4bde95f82506f0e87d0c63d0e04e6ed