Prevent duplicate/unwanted ceph rgw deployments
The reason of this patch is to avoid TripleO to redeploy RGW during FFU,
but in general, this change prevents a redeploy during a stack update in
case the daemon is found.
This means that after day1, an existing RGW deployment can't be updated
at all using TripleO, and should be done via cephadm cli [1].

This patch adds a generic task check_service.yaml to
capture the details of a specific service.

This patch applies same pattern for mds, monitoring stack
and dashboard services.
[1] https://docs.ceph.com/en/quincy/cephadm/operations/

Resolves: rhbz#2229931
Change-Id: Id1e2c2c04f686939824801d3219c21aab4af25bb