Use wallaby tripleo-repos in train job
Train molecule jobs are failing, As mentioned in Closes-Bug:
The tripleo-repos package in RDO train is old and does not
support --stream/--no-stream option. tripleo-repos is branchless,
tripleo-repos is pinned in rdo-info.

Till few month back, we were using master tripleo-repos
in molecule job unintentionally which was discoverred
when master repos become unavailable and was fixed in [2].

With this Patch, we will use centos8-wallaby tripleo-repos
in train jobs. Other alternative is unpinning tripleo-repos
for train branch in rdoinfo, but since train line will be removed
soon. this seems an easy fix here.

[1] https://github.com/redhat-openstack/rdoinfo/blob/8a9bf81c1dc7c132a11bef74d368be3a89fce77f/tags/train.yml#L1195-L1198
[2] https://review.opendev.org/c/openstack/tripleo-ansible/+/885485

Closes-Bug: #2029085
Change-Id: I9f627a6dedcf55c5343d8220b641338905ede74b