Allow setting a custom ExecStopPost command to systemd container units
There are cases where we need or want to use custom ExecStopPost
command, for example to allow the frr service to properly shut down
and close its connections to the peers before re-establishing the
sessions.
This commit allows reading a post_stop_exec attribute from the
container definition so to set the desired value.

Change-Id: Ie485e0772062ff3e3b219be69ac608790ff7baae